調整

+++ src/info/re4k/asfc/aclog/UserDetail.java
@@ -9,10 +9,10 @@ public interface UserDetail{
 	/**
 	 * @return aclog know Favorites count
 	 * */
-	long getFavoritesCount();
+	int getFavoritesCount();
 
 	/**
 	 * @return aclog know Retweets count
 	 * */
-	long getRetweetsCount();
+	int getRetweetsCount();
 }
